requirements.txt
aioredis==1.3.1
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.4.0
django-countries==6.1.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
Pillow==7.1.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.5.3
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
stripe==2.48.0
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.25.9
whitenoise==5.1.0
wrapt==1.12.1
zope.interface==5.1.0
